chore: update get-token-addresses helper
