+++ actions/aave-like/adjust/get-adjust-position-parameters.ts
@@ -1,5 +1,5 @@
 import { AAVETokens, PositionTransition, strategies } from '@oasisdex/dma-library'
-import { getTokenAddresses } from 'actions/aave-like/get-token-addresses'
+import { getAddresses } from 'actions/aave-like/get-addresses'
 import { networkIdToLibraryNetwork, swapCall } from 'actions/aave-like/helpers'
 import { AdjustAaveParameters } from 'actions/aave-like/types'
 import { getRpcProvider } from 'blockchain/networks'
@@ -35,7 +35,7 @@ export async function getAdjustPositionParameters({
     type strategyDependencies = Parameters<typeof strategies.aave.v2.adjust>[1] &
       Parameters<typeof strategies.aave.v3.adjust>[1]
 
-    const addresses = getTokenAddresses(networkId)
+    const addresses = getAddresses(networkId)
 
     const args: strategyArguments = {
       slippage,
