+++ actions/aave-like/types/open-aave-deposit-borrow-parameters.ts
@@ -1,7 +1,7 @@
 import { AAVETokens } from '@oasisdex/dma-library'
 import BigNumber from 'bignumber.js'
 import { ProxyType } from 'features/aave/types'
-import { AaveLendingProtocol } from 'lendingProtocols'
+import { AaveLendingProtocol, AaveLikeLendingProtocol, SparkLendingProtocol } from 'lendingProtocols'
 import { NetworkIds } from '../../../blockchain/networks'
 
 export interface OpenAaveDepositBorrowParameters {
@@ -15,7 +15,7 @@ export interface OpenAaveDepositBorrowParameters {
   userAddress: string
   token?: string
   proxyType: ProxyType
-  protocol: AaveLendingProtocol
+  protocol: AaveLikeLendingProtocol
   networkId: NetworkIds
   positionType: 'Borrow'
 }
