+++ package.json
@@ -45,9 +45,9 @@
     "@metamask/eth-sig-util": "^5.0.2",
     "@next/font": "^13.1.6",
     "@next/mdx": "12.0.4",
-    "@oasisdex/addresses": "0.0.49",
+    "@oasisdex/addresses": "0.0.44",
     "@oasisdex/automation": "^1.5.2",
-    "@oasisdex/dma-library": "0.4.14",
+    "@oasisdex/dma-library": "0.4.10",
     "@oasisdex/multiply": "^0.2.11",
     "@oasisdex/transactions": "0.1.4-alpha.0",
     "@oasisdex/utils": "^0.0.8",
